feat(admin): optimize Superadmin tools, implement cascade suspension, and refactor metrics
- Implemented account suspension validation and cascade suspension (Hacienda Admin -> Operadores) in AuthStore and Vue Router.
- Added "Suspended" to `USER_STATUS` constants and form selects.
- Refactored `SuperAdminDashboard` to use real data widgets instead of static navigation cards.
- Merged `UsageMetrics` into `SuperAdminAnalytics` using tabs, deleting `UsageMetrics.vue`.
- Added new `SuperAdminFinanzas.vue` to inject context and view global financial data.
- Refactored `DataMiningTools` to execute real queries against PocketBase rather than mock data.
