+++ src/components/forms/auth/UserForm.vue
@@ -87,7 +87,7 @@
 </template>
 
 <script setup>
-import { ref, reactive, computed, watch } from 'vue'
+import { ref, reactive, watch } from 'vue'
 
 const props = defineProps({
   initialData: { type: Object, default: () => ({}) },
